src/main/java/com/phonedirectory/controllers/StatusController.java
package com.phonedirectory.controllers;

import lombok.extern.log4j.Log4j;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;

@Log4j
@Controller
public class StatusController {

    @GetMapping(value = "/status", produces = MediaType.TEXT_PLAIN_VALUE)
    public ResponseEntity<String> getStatus() {
        String response = "OK";
        log.info("status of application");

        return ResponseEntity.status(HttpStatus.OK).body(response);
    }

    @DeleteMapping(value = "/exit")
    public void exitingApplication() {
        log.info("exiting application gracefully");

        System.exit(0);
    }
}